Refactor into api version specific packages
