+++ certmanager/v1alpha2/certificate.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package acme
+package v1alpha2
 
 import (
 	"encoding/json"
@@ -22,8 +22,8 @@ import (
 
 	jsonpatch "github.com/evanphx/json-patch"
 	"github.com/golang/glog"
-	api "github.com/jetstack/cert-manager/pkg/apis/acme/v1alpha2"
-	cs "github.com/jetstack/cert-manager/pkg/client/clientset/versioned/typed/acme/v1alpha2"
+	api "github.com/jetstack/cert-manager/pkg/apis/certmanager/v1alpha2"
+	cs "github.com/jetstack/cert-manager/pkg/client/clientset/versioned/typed/certmanager/v1alpha2"
 	"github.com/pkg/errors"
 	kerr "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -32,13 +32,13 @@ import (
 	kutil "kmodules.xyz/client-go"
 )
 
-func CreateOrPatchOrder(c cs.AcmeV1alpha2Interface, meta metav1.ObjectMeta, transform func(alert *api.Order) *api.Order) (*api.Order, kutil.VerbType, error) {
-	cur, err := c.Orders(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
+func CreateOrPatchCertificate(c cs.CertmanagerV1alpha2Interface, meta metav1.ObjectMeta, transform func(alert *api.Certificate) *api.Certificate) (*api.Certificate, kutil.VerbType, error) {
+	cur, err := c.Certificates(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
 	if kerr.IsNotFound(err) {
-		glog.V(3).Infof("Creating Order %s/%s.", meta.Namespace, meta.Name)
-		out, err := c.Orders(meta.Namespace).Create(transform(&api.Order{
+		glog.V(3).Infof("Creating Certificate %s/%s.", meta.Namespace, meta.Name)
+		out, err := c.Certificates(meta.Namespace).Create(transform(&api.Certificate{
 			TypeMeta: metav1.TypeMeta{
-				Kind:       "Order",
+				Kind:       "Certificate",
 				APIVersion: api.SchemeGroupVersion.String(),
 			},
 			ObjectMeta: meta,
@@ -47,14 +47,14 @@ func CreateOrPatchOrder(c cs.AcmeV1alpha2Interface, meta metav1.ObjectMeta, tran
 	} else if err != nil {
 		return nil, kutil.VerbUnchanged, err
 	}
-	return PatchOrder(c, cur, transform)
+	return PatchCertificate(c, cur, transform)
 }
 
-func PatchOrder(c cs.AcmeV1alpha2Interface, cur *api.Order, transform func(*api.Order) *api.Order) (*api.Order, kutil.VerbType, error) {
-	return PatchOrderObject(c, cur, transform(cur.DeepCopy()))
+func PatchCertificate(c cs.CertmanagerV1alpha2Interface, cur *api.Certificate, transform func(*api.Certificate) *api.Certificate) (*api.Certificate, kutil.VerbType, error) {
+	return PatchCertificateObject(c, cur, transform(cur.DeepCopy()))
 }
 
-func PatchOrderObject(c cs.AcmeV1alpha2Interface, cur, mod *api.Order) (*api.Order, kutil.VerbType, error) {
+func PatchCertificateObject(c cs.CertmanagerV1alpha2Interface, cur, mod *api.Certificate) (*api.Certificate, kutil.VerbType, error) {
 	curJson, err := json.Marshal(cur)
 	if err != nil {
 		return nil, kutil.VerbUnchanged, err
@@ -72,39 +72,39 @@ func PatchOrderObject(c cs.AcmeV1alpha2Interface, cur, mod *api.Order) (*api.Ord
 	if len(patch) == 0 || string(patch) == "{}" {
 		return cur, kutil.VerbUnchanged, nil
 	}
-	glog.V(3).Infof("Patching Order %s/%s with %s.", cur.Namespace, cur.Name, string(patch))
-	out, err := c.Orders(cur.Namespace).Patch(cur.Name, types.MergePatchType, patch)
+	glog.V(3).Infof("Patching Certificate %s/%s with %s.", cur.Namespace, cur.Name, string(patch))
+	out, err := c.Certificates(cur.Namespace).Patch(cur.Name, types.MergePatchType, patch)
 	return out, kutil.VerbPatched, err
 }
 
-func TryUpdateOrder(c cs.AcmeV1alpha2Interface, meta metav1.ObjectMeta, transform func(*api.Order) *api.Order) (result *api.Order, err error) {
+func TryUpdateCertificate(c cs.CertmanagerV1alpha2Interface, meta metav1.ObjectMeta, transform func(*api.Certificate) *api.Certificate) (result *api.Certificate, err error) {
 	attempt := 0
 	err = wait.PollImmediate(kutil.RetryInterval, kutil.RetryTimeout, func() (bool, error) {
 		attempt++
-		cur, e2 := c.Orders(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
+		cur, e2 := c.Certificates(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
 		if kerr.IsNotFound(e2) {
 			return false, e2
 		} else if e2 == nil {
-			result, e2 = c.Orders(cur.Namespace).Update(transform(cur.DeepCopy()))
+			result, e2 = c.Certificates(cur.Namespace).Update(transform(cur.DeepCopy()))
 			return e2 == nil, nil
 		}
-		glog.Errorf("Attempt %d failed to update Order %s/%s due to %v.", attempt, cur.Namespace, cur.Name, e2)
+		glog.Errorf("Attempt %d failed to update Certificate %s/%s due to %v.", attempt, cur.Namespace, cur.Name, e2)
 		return false, nil
 	})
 
 	if err != nil {
-		err = errors.Errorf("failed to update Order %s/%s after %d attempts due to %v", meta.Namespace, meta.Name, attempt, err)
+		err = errors.Errorf("failed to update Certificate %s/%s after %d attempts due to %v", meta.Namespace, meta.Name, attempt, err)
 	}
 	return
 }
 
-func UpdateOrderStatus(
-	c cs.AcmeV1alpha2Interface,
-	in *api.Order,
-	transform func(*api.OrderStatus) *api.OrderStatus,
-) (result *api.Order, err error) {
-	apply := func(x *api.Order) *api.Order {
-		return &api.Order{
+func UpdateCertificateStatus(
+	c cs.CertmanagerV1alpha2Interface,
+	in *api.Certificate,
+	transform func(*api.CertificateStatus) *api.CertificateStatus,
+) (result *api.Certificate, err error) {
+	apply := func(x *api.Certificate) *api.Certificate {
+		return &api.Certificate{
 			TypeMeta:   x.TypeMeta,
 			ObjectMeta: x.ObjectMeta,
 			Spec:       x.Spec,
@@ -117,9 +117,9 @@ func UpdateOrderStatus(
 	err = wait.PollImmediate(kutil.RetryInterval, kutil.RetryTimeout, func() (bool, error) {
 		attempt++
 		var e2 error
-		result, e2 = c.Orders(in.Namespace).UpdateStatus(apply(cur))
+		result, e2 = c.Certificates(in.Namespace).UpdateStatus(apply(cur))
 		if kerr.IsConflict(e2) {
-			latest, e3 := c.Orders(in.Namespace).Get(in.Name, metav1.GetOptions{})
+			latest, e3 := c.Certificates(in.Namespace).Get(in.Name, metav1.GetOptions{})
 			switch {
 			case e3 == nil:
 				cur = latest
@@ -136,7 +136,7 @@ func UpdateOrderStatus(
 	})
 
 	if err != nil {
-		err = fmt.Errorf("failed to update status of Order %s/%s after %d attempts due to %v", in.Namespace, in.Name, attempt, err)
+		err = fmt.Errorf("failed to update status of Certificate %s/%s after %d attempts due to %v", in.Namespace, in.Name, attempt, err)
 	}
 	return
 }
