+++ certmanager/v1alpha2/clusterissuer.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package certmanager
+package v1alpha2
 
 import (
 	"encoding/json"
