+++ certmanager/v1alpha2/issuer.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package certmanager
+package v1alpha2
 
 import (
 	"encoding/json"
@@ -32,13 +32,13 @@ import (
 	kutil "kmodules.xyz/client-go"
 )
 
-func CreateOrPatchCertificate(c cs.CertmanagerV1alpha2Interface, meta metav1.ObjectMeta, transform func(alert *api.Certificate) *api.Certificate) (*api.Certificate, kutil.VerbType, error) {
-	cur, err := c.Certificates(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
+func CreateOrPatchIssuer(c cs.CertmanagerV1alpha2Interface, meta metav1.ObjectMeta, transform func(alert *api.Issuer) *api.Issuer) (*api.Issuer, kutil.VerbType, error) {
+	cur, err := c.Issuers(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
 	if kerr.IsNotFound(err) {
-		glog.V(3).Infof("Creating Certificate %s/%s.", meta.Namespace, meta.Name)
-		out, err := c.Certificates(meta.Namespace).Create(transform(&api.Certificate{
+		glog.V(3).Infof("Creating Issuer %s/%s.", meta.Namespace, meta.Name)
+		out, err := c.Issuers(meta.Namespace).Create(transform(&api.Issuer{
 			TypeMeta: metav1.TypeMeta{
-				Kind:       "Certificate",
+				Kind:       "Issuer",
 				APIVersion: api.SchemeGroupVersion.String(),
 			},
 			ObjectMeta: meta,
@@ -47,14 +47,14 @@ func CreateOrPatchCertificate(c cs.CertmanagerV1alpha2Interface, meta metav1.Obj
 	} else if err != nil {
 		return nil, kutil.VerbUnchanged, err
 	}
-	return PatchCertificate(c, cur, transform)
+	return PatchIssuer(c, cur, transform)
 }
 
-func PatchCertificate(c cs.CertmanagerV1alpha2Interface, cur *api.Certificate, transform func(*api.Certificate) *api.Certificate) (*api.Certificate, kutil.VerbType, error) {
-	return PatchCertificateObject(c, cur, transform(cur.DeepCopy()))
+func PatchIssuer(c cs.CertmanagerV1alpha2Interface, cur *api.Issuer, transform func(*api.Issuer) *api.Issuer) (*api.Issuer, kutil.VerbType, error) {
+	return PatchIssuerObject(c, cur, transform(cur.DeepCopy()))
 }
 
-func PatchCertificateObject(c cs.CertmanagerV1alpha2Interface, cur, mod *api.Certificate) (*api.Certificate, kutil.VerbType, error) {
+func PatchIssuerObject(c cs.CertmanagerV1alpha2Interface, cur, mod *api.Issuer) (*api.Issuer, kutil.VerbType, error) {
 	curJson, err := json.Marshal(cur)
 	if err != nil {
 		return nil, kutil.VerbUnchanged, err
@@ -72,39 +72,39 @@ func PatchCertificateObject(c cs.CertmanagerV1alpha2Interface, cur, mod *api.Cer
 	if len(patch) == 0 || string(patch) == "{}" {
 		return cur, kutil.VerbUnchanged, nil
 	}
-	glog.V(3).Infof("Patching Certificate %s/%s with %s.", cur.Namespace, cur.Name, string(patch))
-	out, err := c.Certificates(cur.Namespace).Patch(cur.Name, types.MergePatchType, patch)
+	glog.V(3).Infof("Patching Issuer %s/%s with %s.", cur.Namespace, cur.Name, string(patch))
+	out, err := c.Issuers(cur.Namespace).Patch(cur.Name, types.MergePatchType, patch)
 	return out, kutil.VerbPatched, err
 }
 
-func TryUpdateCertificate(c cs.CertmanagerV1alpha2Interface, meta metav1.ObjectMeta, transform func(*api.Certificate) *api.Certificate) (result *api.Certificate, err error) {
+func TryUpdateIssuer(c cs.CertmanagerV1alpha2Interface, meta metav1.ObjectMeta, transform func(*api.Issuer) *api.Issuer) (result *api.Issuer, err error) {
 	attempt := 0
 	err = wait.PollImmediate(kutil.RetryInterval, kutil.RetryTimeout, func() (bool, error) {
 		attempt++
-		cur, e2 := c.Certificates(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
+		cur, e2 := c.Issuers(meta.Namespace).Get(meta.Name, metav1.GetOptions{})
 		if kerr.IsNotFound(e2) {
 			return false, e2
 		} else if e2 == nil {
-			result, e2 = c.Certificates(cur.Namespace).Update(transform(cur.DeepCopy()))
+			result, e2 = c.Issuers(cur.Namespace).Update(transform(cur.DeepCopy()))
 			return e2 == nil, nil
 		}
-		glog.Errorf("Attempt %d failed to update Certificate %s/%s due to %v.", attempt, cur.Namespace, cur.Name, e2)
+		glog.Errorf("Attempt %d failed to update Issuer %s/%s due to %v.", attempt, cur.Namespace, cur.Name, e2)
 		return false, nil
 	})
 
 	if err != nil {
-		err = errors.Errorf("failed to update Certificate %s/%s after %d attempts due to %v", meta.Namespace, meta.Name, attempt, err)
+		err = errors.Errorf("failed to update Issuer %s/%s after %d attempts due to %v", meta.Namespace, meta.Name, attempt, err)
 	}
 	return
 }
 
-func UpdateCertificateStatus(
+func UpdateIssuerStatus(
 	c cs.CertmanagerV1alpha2Interface,
-	in *api.Certificate,
-	transform func(*api.CertificateStatus) *api.CertificateStatus,
-) (result *api.Certificate, err error) {
-	apply := func(x *api.Certificate) *api.Certificate {
-		return &api.Certificate{
+	in *api.Issuer,
+	transform func(*api.IssuerStatus) *api.IssuerStatus,
+) (result *api.Issuer, err error) {
+	apply := func(x *api.Issuer) *api.Issuer {
+		return &api.Issuer{
 			TypeMeta:   x.TypeMeta,
 			ObjectMeta: x.ObjectMeta,
 			Spec:       x.Spec,
@@ -117,9 +117,9 @@ func UpdateCertificateStatus(
 	err = wait.PollImmediate(kutil.RetryInterval, kutil.RetryTimeout, func() (bool, error) {
 		attempt++
 		var e2 error
-		result, e2 = c.Certificates(in.Namespace).UpdateStatus(apply(cur))
+		result, e2 = c.Issuers(in.Namespace).UpdateStatus(apply(cur))
 		if kerr.IsConflict(e2) {
-			latest, e3 := c.Certificates(in.Namespace).Get(in.Name, metav1.GetOptions{})
+			latest, e3 := c.Issuers(in.Namespace).Get(in.Name, metav1.GetOptions{})
 			switch {
 			case e3 == nil:
 				cur = latest
@@ -136,7 +136,7 @@ func UpdateCertificateStatus(
 	})
 
 	if err != nil {
-		err = fmt.Errorf("failed to update status of Certificate %s/%s after %d attempts due to %v", in.Namespace, in.Name, attempt, err)
+		err = fmt.Errorf("failed to update status of Issuer %s/%s after %d attempts due to %v", in.Namespace, in.Name, attempt, err)
 	}
 	return
 }
